refactor, no done yet
