+++ chi_annotator/__init__.py
@@ -1,1 +1,2 @@
+# init __version__
 __version__ = "0.0.1"
